requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
supabase>=2.5.0
litellm>=1.74.0
redis>=5.0.1
slowapi>=0.1.9
structlog>=23.2.0
httpx>=0.27.0
prometheus-client>=0.19.0
langfuse>=3.2.1
cryptography>=41.0.7
asyncpg>=0.29.0
tenacity>=8.2.3
pybreaker>=1.0.1
sentry-sdk[fastapi]>=1.38.0